daemon: refactor out of cmd into own package
