+++ daemon/log.go
@@ -1,0 +1,7 @@
+package daemon
+
+import (
+	"github.com/tendermint/tendermint/logger"
+)
+
+var log = logger.New("module", "daemon")
